Lots of little bug fixes, moving to new style of signal/slot connections.

+++ mealog/src/settings.cpp
@@ -48,11 +48,11 @@ QList<float> Settings::getDisplayScales() {
 	return DISPLAY_SCALES;
 }
 
-unsigned int Settings::getRefreshInterval() {
+float Settings::getRefreshInterval() {
 	return settings.value("refresh-interval").toFloat();
 }
 
-void Settings::setRefreshInterval(unsigned int interval) {
+void Settings::setRefreshInterval(float interval) {
 	float i = qMin(qMax(interval, MIN_REFRESH_INTERVAL), MAX_REFRESH_INTERVAL);
 	settings.setValue("refresh-interval", QVariant(i));
 }
